refactor(tui): colocate Screen::Help into tui/screens/help.rs
Second screen migration for #287 (Phase 2). Moved into the new
tui/screens/help.rs: handle_key_help, build_help_vm (+ its exclusive
helpers about_topic_lines_plain, help_topic_body, ABOUT_REPO_LINE — same
"only used by this screen" pattern as diff_footer for #289 and
adjust_config_field for the Config migration), render_help_vm,
help_palette_items, plus a HELP_TOPIC const for HelpTopic::for_screen's arm
and a wants_content_scroll predicate for wheel_step's arm (Help is the only
screen so far needing the latter — its topic body scrolls 3 lines/tick but
the topic index is a plain list).

Pure move — bodies and existing tests unchanged. open_help() (keys.rs)
stays: it's a shared cross-screen navigation entry point (top-bar click,
other screens' '?' key), not Help's own dispatch. apply_navigation's and
click_select's Help-specific arms are untouched — per the issue's grilling
session, those are narrow shared multi-screen accessors, not part of the
five registries this issue targets.

Part of #287
