+++ src/tui/render.rs
@@ -56,7 +56,9 @@ fn render_screen_vm(
         super::ScreenVm::Confirm(confirm) => {
             render_confirm_vm(frame, state, confirm, chrome, layout)
         }
-        super::ScreenVm::Help(help) => render_help_vm(frame, state, help, chrome, layout),
+        super::ScreenVm::Help(help) => {
+            super::screens::help::render_help_vm(frame, state, help, chrome, layout)
+        }
         super::ScreenVm::Palette(palette) => {
             render_palette_vm(frame, state, palette, chrome, layout)
         }
@@ -159,326 +161,6 @@ pub(super) fn count_label(shown: usize, total: usize) -> String {
         format!("({shown}/{total})")
     } else {
         format!("({total})")
-    }
-}
-
-pub(super) fn help_topic_body(topic: HelpTopic) -> &'static str {
-    match topic {
-        HelpTopic::List => {
-            "\
-Navigation
-  Tab        switch pane (Local / Gists)
-  1 / 2      jump to the Local / Gist pane
-  Up/Down    move the selection (also j / k)
-  Left/Right scroll a long row horizontally (also h / l)
-  Ctrl+b/f   page up / down by 10 (also PageUp / PageDown)
-
-List screen
-  r          toggle recursive file discovery (skips hidden + configured dirs)
-  /          filter the focused pane (Local = path/filename, Gist = description/id)
-             while filtering: type to match · ↑↓ move · PgUp/PgDn page · Tab apply + switch pane
-             · Enter apply · Esc clear · ←/→/Home/End move · Del
-  v          cycle gist visibility: all / public / secret / starred / forked
-  *          star / unstar the selected gist
-             (others' gists are read-only: preview, diff, download, browser — not pin/upload/delete;
-              open gist detail to fork with F)
-  s          cycle the focused pane's sort: match / name / recent
-  t          toggle row view: description / id
-  T          toggle light/dark colour theme (global; saved to config)
-  a          flip which pane drives match ranking (anchor); the other pane
-             re-ranks against the anchor's selection (focus stays put)
-             (📌 = pinned pair · bold = same filename)
-
-Actions (on the selected local file + gist)
-  Enter      diff the local file against the gist; direction follows the focused
-             pane — Gist pane = download view, Local pane = upload view
-             (--- old / +++ new; local label = yellow, gist label = blue)
-  Space      preview the gist file's content (R in preview to force-refresh;
-             blocked for images/binary — use d to download instead)
-  H          open revision history for the selected gist file
-  d          download the gist into the cwd
-  u          upload the local file into the gist
-  n          create a new gist from the local file (type a description, then s/p)
-  p          pin / unpin the local <-> gist pair
-  P          view / manage all pinned mappings (sync status + s/u/d/x)
-  S          smart-sync the selected pinned pair (push/pull by modified time)
-  X          remove the selected file from its gist (y/n confirm)
-  g          open the gist manager (edit description, delete gist)
-  e          edit the local file in $EDITOR
-  y          copy the selected gist's URL to the system clipboard
-
-Mouse (on by default; disable with mouse = false in config or --no-mouse)
-  Wheel      scroll the focused list or content pane
-  Click      select the clicked row (List panes also switch focus)
-  Dbl-click  open the clicked row (diff / detail / pin diff / preview)
-  Tab click  switch Files / Comments on the Gist details screen
-  [✕] btn    close / go back on any pop-up screen
-  Top bar    click (G)ists / (P)ins / (C)onfig / (?)Help (top-right, every screen)
-  Right-click  open the context menu at the click (same as ;)
-  ; / Ctrl+p   open the menu / command palette from the keyboard (see General)"
-        }
-        HelpTopic::Pins => {
-            "\
-  Up/Down    move between pins (also j / k)
-  PageUp/Dn  page by 10 (also Ctrl+b / Ctrl+f)
-  Left/Right scroll a long local path horizontally (also h / l; ~ = home)
-  /          filter pins by path or filename (↑↓ move · PgUp/PgDn page · Enter apply · Esc clear)
-             ←/→/Home/End move the text cursor · Del deletes ahead
-  o          cycle sort: default / local path / gist filename
-  Enter      diff the selected pair (then d pull / u push from the diff)
-  s          smart-sync (newer side wins; skips if already identical)
-  u          force push  (upload local → gist)
-  d          force pull  (download gist → local, diff + y/n confirm)
-  x          unpin the selected pair
-  status     ✓ synced · ↑ local newer · ↓ remote newer · ✕ missing · ? unknown
-  Each row shows (local <age> · gist <age>) relative modification times."
-        }
-        HelpTopic::GistManager => {
-            "\
-  Up/Down    move between gists (also j / k)
-  PageUp/Dn  page by 10 (also Ctrl+b / Ctrl+f)
-  Left/Right scroll a long description horizontally (also h / l)
-  /          filter gists by description or id (↑↓ move · PgUp/PgDn page · Enter apply · Esc clear)
-  s          cycle sort: updated / created
-  v          cycle visibility: all / public / secret / starred / forked
-  *          star / unstar the selected gist
-  Enter      open the gist detail view (info, file list, comments)
-  o          open the gist in your web browser
-  y          copy the gist's URL to the system clipboard
-  H          open revision history (browse, diff, restore)
-  q / Esc    back to the list
-             (edit description, compact, delete: gist detail only, owned gists)
-  Rows show ☆ N (stargazers), ⑂ N (forks), 💬 N (comments) when non-zero;
-  ★ prefix = you starred it; ⑂ prefix = this gist is a fork."
-        }
-        HelpTopic::GistDetail => {
-            "\
-  Tab        switch tab: Files / Comments (one shows at a time; opens on Files)
-  Up/Down    move the file cursor (Files tab) or scroll comments (also j / k)
-  PageUp/Dn  page comments / file cursor by 10 (also Ctrl+b / Ctrl+f)
-  m          load 30 older comments (Comments tab; also click the top line)
-  Enter      preview the cursor-selected file (file list focused; blocked for binary)
-  1-9        preview the content of the Nth file (full-screen; R refresh, q back)
-             non-text files are tagged (binary) in the list
-  H          open revision history for this gist (target = cursor file)
-  *          star / unstar this gist
-  o          open the gist in your web browser
-  y          copy the gist's URL to the system clipboard
-  q / Esc    back to the gist manager
-  Info line shows ☆ N (stargazers), ⑂ N (forks), 💬 N (comments) when non-zero
-  Owned gists only:
-  e          edit the gist description (Enter apply, Esc cancel)
-             ←/→/Home/End move the text cursor · Del deletes ahead
-  c          compact revisions (y/n confirm; gist info shown as context)
-  X          delete the entire gist and all its files (y/n confirm)
-  Others' gists:
-  F          fork into your account"
-        }
-        HelpTopic::Revisions => {
-            "\
-  Up/Down    move between revisions (also j / k; newest first; row 0 = current)
-  PageUp/Dn  page by 10 (also Ctrl+b / Ctrl+f)
-  Left/Right scroll a long row horizontally (also h / l)
-  Enter      diff this revision vs its parent (incremental; initial = all additions)
-  F          cycle the target file (multi-file gists; wraps)
-  D          diff the target file: selected revision vs current (read-only; no download/upload)
-  r          restore the target file from the selected revision (y/n confirm)
-  q / Esc    back"
-        }
-        HelpTopic::Diff => {
-            "\
-  Up/Down/Left/Right  scroll the diff (also j / k / h / l; Left/Right only when wrap is off)
-  PageUp/Dn  scroll the diff by 10 lines (also Ctrl+b / Ctrl+f)
-  w          toggle soft line wrapping (remembered for the session)
-  c          toggle context: configured radius <-> full file (remembered)
-  d / u      download / upload from the diff
-  syntax     unchanged context lines are syntax-highlighted by file type
-  newline    a file-final-newline-only difference counts as identical
-             (set ignore_trailing_newline = false for byte-exact diffs)
-  Esc / q    back"
-        }
-        HelpTopic::Preview => {
-            "\
-  Up/Down/Left/Right  scroll (also j / k / h / l; Left/Right only when wrap is off)
-  PageUp/Dn  scroll by 10 lines (also Ctrl+b / Ctrl+f)
-  w          toggle soft line wrapping (remembered for the session)
-  y          copy the gist URL · Y copy the file content to the clipboard
-  syntax     known file types are syntax-highlighted
-  R          re-fetch the content
-  Esc / q    back"
-        }
-        HelpTopic::Upload => {
-            "\
-  y          confirm and execute the upload
-  n / Esc    cancel the upload
-  e          edit / redact the upload content in $EDITOR before upload
-             (GUI editors: the diff updates live while the editor stays open;
-             y/e wait until you close it — n still cancels immediately)
-  p          (JSON only) toggle pretty-print formatting
-  s          (JSON only) toggle recursive key sorting"
-        }
-        HelpTopic::Config => {
-            "\
-Settings (C, top-bar (C)onfig, or Ctrl+p → Open settings)
-  Up/Down    move between fields (also j / k)
-  Enter/Space  toggle a boolean, or increase a number
-  h / l      decrease / increase (also ← / →)
-  Esc / q    close (opening Settings never writes config by itself —
-             values are saved only after you change a field)
-
-Fields
-  Theme                  dark / light (also global T)
-  Mouse support          on / off (session still respects --no-mouse)
-  Check for updates      on / off (session still respects --no-update-check)
-  Ignore trailing newline  on / off (diff + overwrite confirm)
-  Recursive scan depth   0–20 (r recursive discovery)
-  Diff context lines     0–50 (c in Diff still toggles full vs this radius)"
-        }
-        HelpTopic::General => {
-            "\
-  Esc / q    close an overlay; from the list, press twice to quit the app
-  ?          show this help
-  C          open Settings (flat list of preferences; also Ctrl+p)
-  ;          context menu (actions valid for the current screen + selection)
-  Ctrl+p     command palette (all actions + cross-screen navigation; type to filter)
-  T          toggle light/dark colour theme (saved to config)
-  Up/Down    scroll this help text
-  NO_COLOR   set this env var to disable syntax highlighting (preview + diff)"
-        }
-        HelpTopic::About => {
-            unreachable!(
-                "About has its own dynamic body in about_topic_lines, rendered before help_topic_body is ever called"
-            )
-        }
-    }
-}
-
-/// Fixed row (0-indexed, within the topic body) of the clickable repo-URL line — used to
-/// place `MouseLayout::repo_link`'s hit-rect. Kept stable regardless of update-check state
-/// (see `about_topic_lines`) so this constant never has to change.
-pub(super) const ABOUT_REPO_LINE: usize = 2;
-
-/// Plain-text About topic lines for the pure view-model (issue #241). Paint re-applies the
-/// underlined repo style from [`ABOUT_REPO_LINE`].
-pub(super) fn about_topic_lines_plain(state: &AppState) -> Vec<String> {
-    let repo = env!("CARGO_PKG_REPOSITORY")
-        .trim_start_matches("https://")
-        .trim_start_matches("http://");
-    let mut lines = vec![
-        format!("gistui v{}", env!("CARGO_PKG_VERSION")),
-        String::new(),
-        format!("  {repo}"),
-        String::new(),
-    ];
-    if let Some(latest) = &state.update_available {
-        lines.push(crate::update_check::update_hint(
-            latest,
-            &state.install_method,
-        ));
-    }
-    lines
-}
-
-fn render_help_vm(
-    frame: &mut Frame,
-    state: &AppState,
-    help: &super::view_model::HelpVm,
-    chrome: &super::view_model::ChromeVm,
-    layout: &mut MouseLayout,
-) {
-    let area = frame.area();
-    let area = render_top_bar(frame, area, &state.theme, chrome.mouse_enabled, layout);
-    match &help.mode {
-        super::view_model::HelpModeVm::Index { items, selected } => {
-            let list_items: Vec<ListItem> = items
-                .iter()
-                .map(|item| ListItem::new(format!("  {}  {}", item.key, item.title)))
-                .collect();
-            let list = List::new(list_items)
-                .block(
-                    Block::default()
-                        .title("Help — pick a topic (1-9,0 / ↑↓ Enter · Esc back)")
-                        .borders(Borders::ALL)
-                        .border_type(BorderType::Rounded)
-                        .style(state.theme.base_style())
-                        .padding(Padding::horizontal(1)),
-                )
-                .style(state.theme.base_style())
-                .highlight_style(
-                    Style::default()
-                        .bg(state.theme.accent)
-                        .fg(state.theme.fg_on_accent)
-                        .add_modifier(Modifier::BOLD),
-                )
-                .highlight_symbol("▶ ");
-            let mut list_state = ListState::default();
-            list_state.select(Some(*selected));
-            frame.render_stateful_widget(list, area, &mut list_state);
-            if chrome.mouse_enabled {
-                layout.list = Some(PaneHit {
-                    rect: area,
-                    offset: list_state.offset(),
-                });
-            }
-        }
-        super::view_model::HelpModeVm::Topic {
-            title,
-            lines,
-            scroll,
-            about_repo_line,
-        } => {
-            let body_lines: Vec<Line<'static>> = lines
-                .iter()
-                .enumerate()
-                .map(|(i, text)| {
-                    if about_repo_line == &Some(i) {
-                        let repo = text.trim_start();
-                        let indent_len = text.len() - repo.len();
-                        let indent = text[..indent_len].to_string();
-                        Line::from(vec![
-                            Span::raw(indent),
-                            Span::styled(
-                                repo.to_string(),
-                                Style::default()
-                                    .fg(state.theme.fg)
-                                    .add_modifier(Modifier::UNDERLINED),
-                            ),
-                        ])
-                    } else {
-                        Line::from(text.clone())
-                    }
-                })
-                .collect();
-            frame.render_widget(
-                Paragraph::new(Text::from(body_lines))
-                    .style(state.theme.base_style())
-                    .scroll((*scroll, 0))
-                    .block(
-                        Block::default()
-                            .title(title.clone())
-                            .borders(Borders::ALL)
-                            .border_type(BorderType::Rounded)
-                            .style(state.theme.base_style())
-                            .padding(Padding::horizontal(1)),
-                    ),
-                area,
-            );
-            if let (Some(repo_line), true) = (*about_repo_line, chrome.mouse_enabled) {
-                let visible_row = repo_line as i32 - *scroll as i32;
-                let inner_height = area.height.saturating_sub(2);
-                if visible_row >= 0 && (visible_row as u16) < inner_height {
-                    layout.repo_link = Some(Rect::new(
-                        area.x + 1,
-                        area.y + 1 + visible_row as u16,
-                        area.width.saturating_sub(2),
-                        1,
-                    ));
-                }
-            }
-        }
-    }
-    if chrome.mouse_enabled {
-        layout.close_button = Some(render_close_button(frame, area, &state.theme));
     }
 }
