+++ src/tui/screens/mod.rs
@@ -5,3 +5,4 @@
 //! keep their exhaustive-match shape; each arm becomes a one-line call into a screen module.
 
 pub(crate) mod config;
+pub(crate) mod help;
